requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn==0.24.0
starlette==0.27.0

# Database and ORM
sqlalchemy==2.0.23
alembic==1.16.1
psycopg2-binary==2.9.10

# Authentication and security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.3.0
python-multipart==0.0.6
cryptography>=41.0.5,<43  # Updated for pyOpenSSL compatibility
cffi==1.17.1

# Enhanced security
python-magic==0.4.27  # For file type detection
aiofiles==23.2.1  # For async file operations
secure==0.3.0  # For security headers
pyOpenSSL==24.0.0  # For SSL/TLS
certifi==2024.2.2  # For SSL certificates
urllib3[secure]==2.2.1  # For secure HTTP
safety==2.3.5  # For dependency security checks
bandit==1.7.8  # For security linting

# Rate limiting
slowapi==0.1.9
redis==5.0.1

# AI and machine learning
openai==1.3.5
tiktoken==0.9.0

# Network automation
netmiko==4.5.0
paramiko==3.5.1
pysnmp==4.4.12
pyasn1==0.4.8
ntc_templates==7.9.0
textfsm==1.1.3

# Data processing and analysis
pandas==2.2.3
ijson==3.4.0
numpy==2.2.6

# Document processing
python-docx==1.0.1
openpyxl==3.1.2
python-magic==0.4.27
pdfplumber==0.11.6
pdfminer.six==20250327
PyPDF2==3.0.1
pypdfium2==4.30.1

# PDF generation
fpdf==1.7.2

# HTTP client
requests==2.32.3
httpx==0.25.1

# Google Cloud services
google-cloud-secret-manager==2.20.0

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.11.5
pydantic-settings==2.1.0
PyYAML==6.0.2
ruamel.yaml==0.18.12
ruamel.yaml.clib==0.2.12

# Testing
pytest==7.4.3
pytest-asyncio==0.23.5  # For async tests
pytest-cov==4.1.0  # For coverage reports

# Timezone handling
pytz==2025.2

# Additional utilities
websockets==15.0.1
watchfiles==1.0.5
rich==14.0.0
tqdm==4.67.1
xlsxwriter==3.2.3 